labeling_tool/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cellpose==4.1.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
einops==0.8.2
fastapi==0.135.2
fastremap==1.18.0
filelock==3.20.0
fill_voids==2.1.1
fire==0.7.1
fonttools==4.62.1
fsspec==2025.12.0
ftfy==6.3.1
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
imagecodecs==2026.3.6
ImageIO==2.37.3
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
monai==1.5.2
mpmath==1.3.0
natsort==8.4.0
networkx==3.6.1
nibabel==5.4.2
numpy==2.3.5
open_clip_torch==3.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.13.0.92
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049096285/work
pillow==12.0.0
polars==1.39.3
polars-runtime-32==1.39.3
protobuf==7.34.1
psutil==7.2.2
pybboxes==0.1.6
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.3.32
requests==2.32.5
rich==14.3.3
roifile==2026.2.10
safetensors==0.7.0
sahi==0.11.36
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
segment-anything==1.0
shapely==2.1.2
shellingham==1.5.4
simpleitk==2.5.3
six==1.17.0
starlette==1.0.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
terminaltables==3.1.10
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.26
torch==2.10.0+cu128
torchaudio==2.10.0+cu128
torchvision==0.25.0+cu128
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.24
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
Werkzeug==3.1.7
